LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource-timing-level1.sub-expected.txt


PASS No timeline entry for about:blank
PASS 'iframe (Populate cache): The initial request populates the cache (if appropriate).
PASS 'iframe (Potentially Cached): Immediately fetch the same URL, exercising the cache hit path (if any).
PASS 'xmlhttprequest (Populate cache): The initial request populates the cache (if appropriate).
PASS 'xmlhttprequest (Potentially Cached): Immediately fetch the same URL, exercising the cache hit path (if any).
PASS 'script (Populate cache): The initial request populates the cache (if appropriate).
PASS 'script (Potentially Cached): Immediately fetch the same URL, exercising the cache hit path (if any).
PASS 'link (Populate cache): The initial request populates the cache (if appropriate).
PASS 'link (Potentially Cached): Immediately fetch the same URL, exercising the cache hit path (if any).
PASS 'iframe: 250ms delay before 'responseStart', another 250ms delay before 'responseEnd'.
FAIL 'xmlhttprequest: 250ms delay before 'responseStart', another 250ms delay before 'responseEnd'. assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'script: 250ms delay before 'responseStart', another 250ms delay before 'responseEnd'. assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'link: 250ms delay before 'responseStart', another 250ms delay before 'responseEnd'. assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'iframe (Redirected): 250ms delay before 'redirectEnd', another 250ms delay before 'responseStart'. assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'xmlhttprequest (Redirected): 250ms delay before 'redirectEnd', another 250ms delay before 'responseStart'. assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'script (Redirected): 250ms delay before 'redirectEnd', another 250ms delay before 'responseStart'. assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'link (Redirected): 250ms delay before 'redirectEnd', another 250ms delay before 'responseStart'. assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'iframe 250ms delay in headers does not affect responseStart' assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'xmlhttprequest 250ms delay in headers does not affect responseStart' assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'script 250ms delay in headers does not affect responseStart' assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'link 250ms delay in headers does not affect responseStart' assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'iframe responseStart uses 1XX (first) response timings' assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'xmlhttprequest responseStart uses 1XX (first) response timings' assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'script responseStart uses 1XX (first) response timings' assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
FAIL 'link responseStart uses 1XX (first) response timings' assert_less_than_equal: connectEnd <= requestStart expected a number less than or equal to 2792.0000000000005 but got 2793
Show details
